requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
marshmallow==3.20.1
werkzeug==2.3.7